Finishing up UserRespository and Tests
Completes the funcitonality of UserRepository and gets all the tests
written. Also refactored Database and DatabaseCreator a bit to make it
easier to unit test without the tests creating data for each other.

+++ PTS.Entity/Domain/User.cs
@@ -1,7 +1,7 @@
 namespace PTS.Entity.Domain;
 
 public class User {
-    public int Id {get; set;}
+    public long Id {get; set;}
 
     public string Username {get; set;}
 
@@ -13,7 +13,7 @@ public class User {
     
     public DateTime UpdatedAt {get; set;}
 
-    public int PasswordHashVersion {get; set;}
+    public long PasswordHashVersion {get; set;}
 
     public string PasswordHash {get; set;}
